Rename Mocks folders in Fakes
