+++ server/tests/rights_access_context/unit/use_cases/test_grant_access_use_case.py
@@ -4,7 +4,7 @@ from uuid import UUID
 from rights_access_context.application.use_cases import (
     GrantAccessUseCase,
 )
-from ..mocks import FakeRightsRepository
+from ..fakes import FakeRightsRepository
 
 
 @pytest.fixture()
