+++ server/tests/vault_management_context/unit/conftest.py
@@ -3,7 +3,7 @@ import pytest
 from vault_management_context.adapters.secondary.gateways import (
     InMemoryVaultSessionGateway,
 )
-from .mocks import (
+from .fakes import (
     FakeVaultRepository,
     FakeShamirGateway,
     FakeEncryptionGateway,
